cmd/glue: split the main.go monolith into per-subcommand files (closes #330)
Pure mechanical re-org of package main — declaration movement only, no
behavior change (verified: sorted func/type signatures identical to HEAD):

- main.go (396 lines): entry point, subcommand dispatch, usage/version
  text, env-file loading, provider resolution, newAgent, shared flag types.
- run.go: glue run — one-shot/scripted/TUI runs, coding-tool assembly,
  local permission prompting.
- usage.go: token-usage summary + pricing, shared by run and connect.
- serve.go: glue serve — daemon serving, token + metadata helpers.
- connect.go: glue connect — config resolution, run start/SSE streaming,
  permission brokering, cancel.
- connect_inspect.go: read-only connect surfaces — status, diagnose, and
  the daemon catalogs with their text renderers.
- main_test.go split along the same seams into run_test.go,
  serve_test.go, connect_test.go; shared fakes stay in main_test.go.
