+++ cmd/glue/main.go
@@ -10,15 +10,10 @@ package main
 
 import (
 	"bufio"
-	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
-	"flag"
 	"fmt"
 	"io"
 	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
 	"runtime"
@@ -27,17 +22,11 @@ import (
 	"syscall"
 	"time"
 
-	"sort"
-
 	"golang.org/x/term"
 
 	"github.com/erain/glue"
-	"github.com/erain/glue/cmd/glue/atmentions"
-	"github.com/erain/glue/cmd/glue/tui"
-	"github.com/erain/glue/daemon"
 	"github.com/erain/glue/providers"
 	filestore "github.com/erain/glue/stores/file"
-	toolscoding "github.com/erain/glue/tools/coding"
 
 	// Register the shipped providers so they resolve through the
 	// providers registry by name (--provider). Importing for side
@@ -49,7 +38,9 @@ import (
 )
 
 const defaultProvider = "gemini"
+
 const defaultListenAddr = "127.0.0.1:0"
+
 const defaultShutdownTimeout = 5 * time.Second
 
 // providerFactory constructs a [glue.Provider] for the named provider, or
@@ -74,178 +65,6 @@ func (r *repeatedStrings) String() string { return strings.Join(*r, ",") }
 func (r *repeatedStrings) Set(value string) error {
 	*r = append(*r, value)
 	return nil
-}
-
-type codingFlagConfig struct {
-	Enabled         bool
-	WorkDir         string
-	AllowedBinaries []string
-	AllowOverwrite  bool
-}
-
-type agentConfig struct {
-	Provider   string
-	Model      string
-	StoreDir   string
-	WorkDir    string
-	Tools      []glue.Tool
-	Permission glue.Permission
-}
-
-type serveFunc func(context.Context, serveConfig, http.Handler, io.Writer) error
-
-type serveConfig struct {
-	ListenAddr        string
-	Token             string
-	TokenSource       string
-	MetadataPath      string
-	Model             string
-	StoreDir          string
-	WorkDir           string
-	PermissionTimeout time.Duration
-	ShutdownTimeout   time.Duration
-}
-
-type daemonMetadata = daemon.Metadata
-
-type httpDoer interface {
-	Do(*http.Request) (*http.Response, error)
-}
-
-type connectConfig struct {
-	BaseURL      string
-	Token        string
-	MetadataPath string
-	SessionID    string
-	Prompt       string
-	Skill        string
-	SkillArgs    map[string]string
-	ClientID     string
-	Model        string
-	Role         string
-	MaxTurns     int
-}
-
-type startRunPayload struct {
-	Text      string            `json:"text,omitempty"`
-	Skill     string            `json:"skill,omitempty"`
-	Arguments map[string]string `json:"arguments,omitempty"`
-	ClientID  string            `json:"client_id,omitempty"`
-	Role      string            `json:"role,omitempty"`
-	Model     string            `json:"model,omitempty"`
-	MaxTurns  int               `json:"max_turns,omitempty"`
-}
-
-type startRunResult struct {
-	RunID     string `json:"run_id"`
-	SessionID string `json:"session_id"`
-	EventsURL string `json:"events_url"`
-}
-
-type connectRunDonePayload struct {
-	Text        string         `json:"text"`
-	Message     *glue.Message  `json:"message,omitempty"`
-	NewMessages []glue.Message `json:"new_messages,omitempty"`
-}
-
-type usageSummary struct {
-	HasUsage         bool
-	InputTokens      int64
-	OutputTokens     int64
-	CacheReadTokens  int64
-	CacheWriteTokens int64
-	TotalTokens      int64
-}
-
-type usagePricing struct {
-	Enabled                 bool
-	InputUSDPerMillion      float64
-	OutputUSDPerMillion     float64
-	CacheReadUSDPerMillion  float64
-	CacheWriteUSDPerMillion float64
-}
-
-type daemonToolCatalog struct {
-	Tools []daemonToolCatalogEntry `json:"tools"`
-}
-
-type daemonSkillCatalog struct {
-	Skills []daemon.SkillCatalogEntry `json:"skills"`
-}
-
-type daemonRoleCatalog struct {
-	Roles []daemon.RoleCatalogEntry `json:"roles"`
-}
-
-type daemonMCPResourceCatalog struct {
-	Resources []daemon.MCPResourceCatalogEntry `json:"resources"`
-}
-
-type daemonMCPPromptCatalog struct {
-	Prompts []daemon.MCPPromptCatalogEntry `json:"prompts"`
-}
-
-type daemonPermissionCatalog struct {
-	Permissions []daemon.PermissionGrant `json:"permissions"`
-}
-
-type daemonToolCatalogEntry struct {
-	Name                    string          `json:"name"`
-	Description             string          `json:"description,omitempty"`
-	Parameters              json.RawMessage `json:"parameters,omitempty"`
-	RequiresPermission      bool            `json:"requires_permission"`
-	PermissionAction        string          `json:"permission_action,omitempty"`
-	PermissionTargetPreview string          `json:"permission_target_preview,omitempty"`
-}
-
-type daemonStatus struct {
-	OK           bool     `json:"ok"`
-	Version      int      `json:"version"`
-	ActiveRuns   int      `json:"active_runs"`
-	ToolsCount   int      `json:"tools_count"`
-	Capabilities []string `json:"capabilities"`
-}
-
-type daemonDiagnostics = daemon.DiagnosticResponse
-
-type daemonDiagnosis struct {
-	OK            bool               `json:"ok"`
-	State         string             `json:"state"`
-	Summary       string             `json:"summary"`
-	MetadataPath  string             `json:"metadata_path,omitempty"`
-	MetadataFound bool               `json:"metadata_found"`
-	MetadataError string             `json:"metadata_error,omitempty"`
-	MetadataPID   int                `json:"metadata_pid,omitempty"`
-	BaseURL       string             `json:"base_url,omitempty"`
-	BaseURLSource string             `json:"base_url_source,omitempty"`
-	TokenSource   string             `json:"token_source,omitempty"`
-	HTTPStatus    int                `json:"http_status,omitempty"`
-	Status        *daemonStatus      `json:"status,omitempty"`
-	Diagnostics   *daemonDiagnostics `json:"diagnostics,omitempty"`
-	Suggestions   []string           `json:"suggestions,omitempty"`
-}
-
-type daemonInspect struct {
-	Status       daemonStatus                     `json:"status"`
-	Tools        []daemonToolCatalogEntry         `json:"tools"`
-	Skills       []daemon.SkillCatalogEntry       `json:"skills,omitempty"`
-	Roles        []daemon.RoleCatalogEntry        `json:"roles,omitempty"`
-	Memories     []daemon.MemoryEntry             `json:"memories,omitempty"`
-	Permissions  []daemon.PermissionGrant         `json:"permissions,omitempty"`
-	MCPResources []daemon.MCPResourceCatalogEntry `json:"mcp_resources,omitempty"`
-	MCPPrompts   []daemon.MCPPromptCatalogEntry   `json:"mcp_prompts,omitempty"`
-}
-
-type connectPermissionPayload struct {
-	PermissionID string                 `json:"permission_id"`
-	Request      glue.PermissionRequest `json:"request"`
-	ExpiresAt    time.Time              `json:"expires_at"`
-}
-
-type connectPermissionDecision struct {
-	Allow       bool   `json:"allow"`
-	Reason      string `json:"reason,omitempty"`
-	RememberFor string `json:"remember_for,omitempty"`
 }
 
 func main() {
@@ -320,2354 +139,6 @@ func runCLIWithDeps(ctx context.Context, args []string, stdin io.Reader, stdout 
 	}
 }
 
-func runCommand(ctx context.Context, args []string, stdin io.Reader, stdout io.Writer, stderr io.Writer, newProvider providerFactory) error {
-	flags := flag.NewFlagSet("glue run", flag.ContinueOnError)
-	flags.SetOutput(stderr)
-
-	id := flags.String("id", "default", "session id")
-	prompt := flags.String("prompt", "", "prompt text")
-	provider := flags.String("provider", defaultProvider, "provider name: codex, gemini, nvidia, or openrouter")
-	model := flags.String("model", "", "model id (default: the provider's default model); gemini/<model> accepted")
-	storeDir := flags.String("store", ".glue/sessions", "session store directory")
-	workDir := flags.String("work", ".", "working directory for AGENTS.md, skills, roles, and optional coding tools")
-	coding := flags.Bool("coding", false, "enable local coding tools for this run")
-	codingAllowOverwrite := flags.Bool("coding-allow-overwrite", false, "allow write_file to replace existing files after model and permission approval")
-	yolo := flags.Bool("yolo", false, "auto-allow all side-effecting tool calls (write_file / edit_file / shell_exec / mcp). implies --coding-allow-overwrite. always use on a feature branch.")
-	showUsage := flags.Bool("usage", false, "print token usage summary to stderr when available")
-	mode := flags.String("mode", "text", `output mode for one-shot runs: "text" (default streamed text) or "json" (JSON-Lines events)`)
-	toolsAllow := flags.String("tools", "", "comma-separated allowlist of tool names to register (empty = all configured)")
-	noTools := flags.Bool("no-tools", false, "register zero tools (read-only/text-only run)")
-	usagePricing := registerUsagePricingFlags(flags)
-	var envs envFiles
-	flags.Var(&envs, "env", "env file path; repeatable")
-	var allowedBinaries repeatedStrings
-	flags.Var(&allowedBinaries, "allow-binary", "allowed shell_exec binary basename for --coding; repeatable")
-
-	if err := flags.Parse(args); err != nil {
-		return err
-	}
-	markUsagePricingFlagState(flags, usagePricing)
-	if err := validateUsagePricing(*usagePricing); err != nil {
-		return err
-	}
-
-	agentName := "default"
-	if flags.NArg() > 0 {
-		agentName = flags.Arg(0)
-	}
-	if agentName != "default" && agentName != "gemini" {
-		return fmt.Errorf("unknown agent %q; only 'default' is available in this runner", agentName)
-	}
-
-	// Resolve the prompt and the run mode:
-	//   - --prompt set                            → one-shot (today's behavior, unchanged)
-	//   - --prompt empty + stdin not a TTY        → read stdin as the prompt, one-shot
-	//   - --prompt empty + stdin AND stdout TTYs  → interactive TUI
-	//   - --prompt empty + only stdin TTY         → refuse (no place to draw)
-	effectivePrompt := strings.TrimSpace(*prompt)
-	interactive := false
-	if effectivePrompt == "" {
-		stdinTTY := isFileTTY(stdin)
-		stdoutTTY := isFileTTY(stdout)
-		switch {
-		case !stdinTTY:
-			// Piped or test input. Read everything as the prompt.
-			data, err := io.ReadAll(stdin)
-			if err != nil {
-				return fmt.Errorf("read prompt from stdin: %w", err)
-			}
-			effectivePrompt = strings.TrimSpace(string(data))
-			if effectivePrompt == "" {
-				return errors.New("missing required --prompt (and stdin was empty)")
-			}
-		case stdinTTY && stdoutTTY:
-			interactive = true
-		default:
-			return errors.New("missing required --prompt (stdout is not a terminal; interactive mode unavailable)")
-		}
-	}
-
-	providerName, effectiveModel, err := resolveProvider(*provider, *model)
-	if err != nil {
-		return err
-	}
-	// Validate output mode. The TUI ignores --mode (it has its own
-	// rendering); --mode json applies to one-shot text-streaming runs.
-	switch strings.ToLower(*mode) {
-	case "", "text", "json":
-		// ok
-	default:
-		return fmt.Errorf("unknown --mode %q (want text or json)", *mode)
-	}
-	if interactive && strings.ToLower(*mode) == "json" {
-		return errors.New("--mode json is only valid with --prompt or piped stdin (interactive TUI has its own output)")
-	}
-	jsonMode := strings.EqualFold(*mode, "json")
-
-	if err := loadEnvFiles(envs); err != nil {
-		return err
-	}
-
-	// --yolo upgrades the overwrite policy too: the user explicitly told
-	// the model "do whatever," so making write_file ask permission to
-	// overwrite would be incoherent.
-	effectiveAllowOverwrite := *codingAllowOverwrite
-	if *yolo {
-		effectiveAllowOverwrite = true
-	}
-
-	tools, _, err := buildCodingTools(codingFlagConfig{
-		Enabled:         *coding,
-		WorkDir:         *workDir,
-		AllowedBinaries: append([]string(nil), allowedBinaries...),
-		AllowOverwrite:  effectiveAllowOverwrite,
-	})
-	if err != nil {
-		return err
-	}
-	// Filter the tool set down: --no-tools strips them all (a "text-only"
-	// run with no agentic capabilities); --tools is an allowlist by name.
-	// Both apply on top of --coding so the user can say
-	// `--coding --tools read_file,grep` for a read-only coding agent.
-	tools, err = filterTools(tools, *toolsAllow, *noTools)
-	if err != nil {
-		return err
-	}
-	// @file expansion happens after env load (so any path side-effects
-	// are honored) and before agent construction. Skipped mentions are
-	// reported to stderr; the user's original "@path" stays in the
-	// prompt verbatim so the model can see they asked.
-	atRes, err := atmentions.Expand(effectivePrompt, atmentions.Options{WorkDir: *workDir})
-	if err != nil {
-		return fmt.Errorf("expand @-mentions: %w", err)
-	}
-	effectivePrompt = atRes.Prompt
-	for _, skip := range atRes.Skipped {
-		fmt.Fprintf(stderr, "glue run: %s: %s\n", skip.Mention, skip.Reason)
-	}
-
-	// In interactive mode the TUI installs its own permission bridge
-	// per-prompt; the agent-level Permission is left nil so the bridge
-	// gets to render the prompt instead of fighting with the readline
-	// version. --yolo overrides both paths: every side-effecting tool
-	// call is auto-approved without a prompt round-trip.
-	var permission glue.Permission
-	switch {
-	case *yolo:
-		permission = yoloPermission{}
-		fmt.Fprintf(stderr, "glue run: --yolo enabled; permission prompts are off (use on a feature branch).\n")
-	case *coding && !interactive:
-		permission = newLocalPromptPermission(stdin, stderr)
-		fmt.Fprintf(stderr, "glue run: coding tools enabled for %s\n", *workDir)
-	}
-
-	// Construct the provider and store once so the TUI can reuse them
-	// for /compact (needs a Provider for SummarizingCompactor) and
-	// /resume (needs the Store to list sessions). The agent then takes
-	// the same pair via AgentOptions.
-	providerImpl, err := newProvider(providerName)
-	if err != nil {
-		return err
-	}
-	storeImpl := filestore.New(*storeDir)
-	agent := glue.NewAgent(glue.AgentOptions{
-		Provider:   providerImpl,
-		Model:      normalizeModel(effectiveModel),
-		Tools:      append([]glue.Tool(nil), tools...),
-		Store:      storeImpl,
-		WorkDir:    *workDir,
-		Permission: permission,
-	})
-
-	if interactive {
-		// buildToolsAt re-roots the run's coding tool set for /goal -w
-		// worktree isolation, honoring the same flags as the main set.
-		// Only offered when --coding is on: an isolated goal without
-		// coding tools could not touch its worktree anyway.
-		var buildToolsAt func(dir string) ([]glue.Tool, error)
-		if *coding {
-			buildToolsAt = func(dir string) ([]glue.Tool, error) {
-				wtTools, _, err := buildCodingTools(codingFlagConfig{
-					Enabled:         true,
-					WorkDir:         dir,
-					AllowedBinaries: append([]string(nil), allowedBinaries...),
-					AllowOverwrite:  effectiveAllowOverwrite,
-				})
-				if err != nil {
-					return nil, err
-				}
-				return filterTools(wtTools, *toolsAllow, *noTools)
-			}
-		}
-		return tui.Run(ctx, tui.Config{
-			Agent:        agent,
-			SessionID:    *id,
-			Provider:     providerName,
-			Model:        effectiveModel,
-			WorkDir:      *workDir,
-			Tools:        tools,
-			Store:        storeImpl,
-			ProviderImpl: providerImpl,
-			AlwaysAllow:  *yolo,
-			BuildTools:   buildToolsAt,
-		})
-	}
-	session, err := agent.Session(ctx, *id)
-	if err != nil {
-		return err
-	}
-
-	if jsonMode {
-		return runOneShotJSON(ctx, session, effectivePrompt, *id, providerName, effectiveModel, stdout, *showUsage, usagePricing, stderr)
-	}
-
-	wroteDelta := false
-	response, err := session.Prompt(ctx, effectivePrompt, glue.WithEvents(func(event glue.Event) {
-		if event.Type == glue.EventTextDelta && event.Delta != "" {
-			fmt.Fprint(stdout, event.Delta)
-			wroteDelta = true
-		}
-	}))
-	if err != nil {
-		if wroteDelta {
-			fmt.Fprintln(stdout)
-		}
-		return err
-	}
-	if wroteDelta {
-		fmt.Fprintln(stdout)
-	} else if response.Text != "" {
-		fmt.Fprintln(stdout, response.Text)
-	}
-	if *showUsage {
-		writeUsageSummary(stderr, summarizeUsage(response.NewMessages), *usagePricing)
-	}
-	return nil
-}
-
-// runOneShotJSON streams loop events to stdout as JSON Lines. Event
-// shape is intentionally small and stable across releases so scripts
-// (editors, CI, /codeagent integrations) can consume it:
-//
-//	{"type":"start","session":"...","provider":"...","model":"..."}
-//	{"type":"text","delta":"..."}
-//	{"type":"tool_start","call_id":"...","name":"...","args":{...}}
-//	{"type":"tool_end","call_id":"...","is_error":false,"text":"..."}
-//	{"type":"done","text":"...","stop_reason":"..."}
-//	{"type":"error","message":"..."}
-func runOneShotJSON(
-	ctx context.Context,
-	session *glue.Session,
-	prompt, sessionID, providerName, modelName string,
-	stdout io.Writer,
-	showUsage bool,
-	pricing *usagePricing,
-	stderr io.Writer,
-) error {
-	enc := json.NewEncoder(stdout)
-	emit := func(event map[string]any) {
-		_ = enc.Encode(event)
-	}
-	emit(map[string]any{
-		"type":     "start",
-		"session":  sessionID,
-		"provider": providerName,
-		"model":    modelName,
-	})
-	response, err := session.Prompt(ctx, prompt, glue.WithEvents(func(event glue.Event) {
-		switch event.Type {
-		case glue.EventTextDelta:
-			if event.Delta == "" {
-				return
-			}
-			emit(map[string]any{"type": "text", "delta": event.Delta})
-		case glue.EventToolStart:
-			if event.ToolCall == nil {
-				return
-			}
-			var args any
-			if len(event.ToolCall.Arguments) > 0 {
-				_ = json.Unmarshal(event.ToolCall.Arguments, &args)
-			}
-			emit(map[string]any{
-				"type":    "tool_start",
-				"call_id": event.ToolCall.ID,
-				"name":    event.ToolCall.Name,
-				"args":    args,
-			})
-		case glue.EventToolEnd:
-			text := ""
-			isErr := false
-			if event.ToolResult != nil {
-				isErr = event.ToolResult.IsError
-				var parts []string
-				for _, p := range event.ToolResult.Content {
-					if p.Type == glue.ContentTypeText && p.Text != "" {
-						parts = append(parts, p.Text)
-					}
-				}
-				text = strings.Join(parts, "\n")
-			}
-			emit(map[string]any{
-				"type":     "tool_end",
-				"call_id":  event.ToolCallID,
-				"is_error": isErr,
-				"text":     text,
-			})
-		case glue.EventError:
-			if event.Error != "" {
-				emit(map[string]any{"type": "error", "message": event.Error})
-			}
-		}
-	}))
-	if err != nil {
-		emit(map[string]any{"type": "error", "message": err.Error()})
-		return err
-	}
-	stopReason := ""
-	if len(response.NewMessages) > 0 {
-		last := response.NewMessages[len(response.NewMessages)-1]
-		stopReason = string(last.StopReason)
-	}
-	emit(map[string]any{
-		"type":        "done",
-		"text":        response.Text,
-		"stop_reason": stopReason,
-	})
-	if showUsage {
-		writeUsageSummary(stderr, summarizeUsage(response.NewMessages), *pricing)
-	}
-	return nil
-}
-
-// filterTools applies --tools / --no-tools to the configured tool set.
-func filterTools(in []glue.Tool, allowCSV string, noTools bool) ([]glue.Tool, error) {
-	if noTools {
-		if strings.TrimSpace(allowCSV) != "" {
-			return nil, errors.New("--tools and --no-tools are mutually exclusive")
-		}
-		return nil, nil
-	}
-	allowCSV = strings.TrimSpace(allowCSV)
-	if allowCSV == "" {
-		return in, nil
-	}
-	want := map[string]bool{}
-	for _, raw := range strings.Split(allowCSV, ",") {
-		name := strings.TrimSpace(raw)
-		if name == "" {
-			continue
-		}
-		want[name] = true
-	}
-	if len(want) == 0 {
-		return nil, nil
-	}
-	var out []glue.Tool
-	seen := map[string]bool{}
-	for _, t := range in {
-		if want[t.Name] {
-			out = append(out, t)
-			seen[t.Name] = true
-		}
-	}
-	// Refuse unknown names so a typo doesn't silently produce zero tools.
-	var missing []string
-	for name := range want {
-		if !seen[name] {
-			missing = append(missing, name)
-		}
-	}
-	if len(missing) > 0 {
-		sort.Strings(missing)
-		available := make([]string, 0, len(in))
-		for _, t := range in {
-			available = append(available, t.Name)
-		}
-		sort.Strings(available)
-		return nil, fmt.Errorf("--tools: unknown tool name(s) %v (available: %v)", missing, available)
-	}
-	return out, nil
-}
-
-func serveCommand(ctx context.Context, args []string, stdout io.Writer, stderr io.Writer, newProvider providerFactory, serve serveFunc) error {
-	flags := flag.NewFlagSet("glue serve", flag.ContinueOnError)
-	flags.SetOutput(stderr)
-
-	provider := flags.String("provider", defaultProvider, "provider name: codex, gemini, nvidia, or openrouter")
-	model := flags.String("model", "", "model id (default: the provider's default model); gemini/<model> accepted")
-	storeDir := flags.String("store", ".glue/sessions", "session store directory")
-	workDir := flags.String("work", ".", "working directory for AGENTS.md, skills, and roles")
-	coding := flags.Bool("coding", false, "enable local coding tools for daemon runs")
-	codingAllowOverwrite := flags.Bool("coding-allow-overwrite", false, "allow write_file to replace existing files after model and daemon permission approval")
-	listenAddr := flags.String("listen", defaultListenAddr, "local listen address")
-	tokenFlag := flags.String("token", "", "bearer token; defaults to GLUE_DAEMON_TOKEN or a generated token")
-	metadataPath := flags.String("metadata", defaultMetadataPath(), "connection metadata JSON path; empty disables metadata file")
-	permissionTimeout := flags.Duration("permission-timeout", 0, "permission decision timeout; 0 uses daemon default")
-	var envs envFiles
-	flags.Var(&envs, "env", "env file path; repeatable")
-	var allowedBinaries repeatedStrings
-	flags.Var(&allowedBinaries, "allow-binary", "allowed shell_exec binary basename for --coding; repeatable")
-
-	if err := flags.Parse(args); err != nil {
-		return err
-	}
-
-	agentName := "default"
-	if flags.NArg() > 0 {
-		agentName = flags.Arg(0)
-	}
-	if agentName != "default" && agentName != "gemini" {
-		return fmt.Errorf("unknown agent %q; only 'default' is available in this runner", agentName)
-	}
-	providerName, effectiveModel, err := resolveProvider(*provider, *model)
-	if err != nil {
-		return err
-	}
-	if err := loadEnvFiles(envs); err != nil {
-		return err
-	}
-	token, tokenSource, err := resolveDaemonToken(*tokenFlag)
-	if err != nil {
-		return err
-	}
-	if strings.TrimSpace(*metadataPath) == "" && tokenSource == "generated" {
-		return errors.New("metadata disabled requires --token or GLUE_DAEMON_TOKEN")
-	}
-	tools, _, err := buildCodingTools(codingFlagConfig{
-		Enabled:         *coding,
-		WorkDir:         *workDir,
-		AllowedBinaries: append([]string(nil), allowedBinaries...),
-		AllowOverwrite:  *codingAllowOverwrite,
-	})
-	if err != nil {
-		return err
-	}
-	if *coding {
-		fmt.Fprintf(stderr, "glue serve: coding tools enabled for %s\n", *workDir)
-	}
-	agent, err := newAgent(newProvider, agentConfig{
-		Provider: providerName,
-		Model:    effectiveModel,
-		StoreDir: *storeDir,
-		WorkDir:  *workDir,
-		Tools:    tools,
-	})
-	if err != nil {
-		return err
-	}
-	handler, err := daemon.New(daemon.Options{
-		Host:  agent,
-		Token: token,
-		Diagnostics: daemon.DiagnosticInfo{
-			Name:         "glue",
-			ListenAddr:   *listenAddr,
-			MetadataPath: *metadataPath,
-			TokenSource:  tokenSource,
-			Provider:     providerName,
-			Model:        normalizeModel(effectiveModel),
-			StoreType:    "file",
-			StorePath:    *storeDir,
-		},
-		PermissionTimeout: *permissionTimeout,
-	})
-	if err != nil {
-		return err
-	}
-	return serve(ctx, serveConfig{
-		ListenAddr:        *listenAddr,
-		Token:             token,
-		TokenSource:       tokenSource,
-		MetadataPath:      *metadataPath,
-		Model:             normalizeModel(effectiveModel),
-		StoreDir:          *storeDir,
-		WorkDir:           *workDir,
-		PermissionTimeout: *permissionTimeout,
-		ShutdownTimeout:   defaultShutdownTimeout,
-	}, handler, stdout)
-}
-
-func connectCommand(ctx context.Context, args []string, stdin io.Reader, stdout io.Writer, stderr io.Writer, client httpDoer) error {
-	flags := flag.NewFlagSet("glue connect", flag.ContinueOnError)
-	flags.SetOutput(stderr)
-
-	sessionID := flags.String("id", "default", "session id")
-	prompt := flags.String("prompt", "", "prompt text")
-	skillName := flags.String("skill", "", "daemon skill name to run instead of --prompt")
-	baseURL := flags.String("base-url", "", "daemon base URL; defaults to metadata file")
-	tokenFlag := flags.String("token", "", "daemon bearer token; defaults to metadata file or GLUE_DAEMON_TOKEN")
-	metadataPath := flags.String("metadata", defaultMetadataPath(), "connection metadata JSON path; empty disables metadata file")
-	clientID := flags.String("client-id", defaultClientID(), "daemon client id")
-	model := flags.String("model", "", "per-run model override")
-	role := flags.String("role", "", "per-run role override")
-	maxTurns := flags.Int("max-turns", 0, "per-run loop turn budget")
-	showUsage := flags.Bool("usage", false, "print token usage summary to stderr when available")
-	usagePricing := registerUsagePricingFlags(flags)
-	showTools := flags.Bool("tools", false, "list daemon tools and exit without starting a run")
-	toolsJSON := flags.Bool("tools-json", false, "print --tools output as JSON")
-	showSkills := flags.Bool("skills", false, "list daemon skills and exit without starting a run")
-	skillsJSON := flags.Bool("skills-json", false, "print --skills output as JSON")
-	showRoles := flags.Bool("roles", false, "list daemon roles and exit without starting a run")
-	rolesJSON := flags.Bool("roles-json", false, "print --roles output as JSON")
-	showMCPResources := flags.Bool("mcp-resources", false, "list daemon MCP resources and exit without starting a run")
-	mcpResourcesJSON := flags.Bool("mcp-resources-json", false, "print --mcp-resources output as JSON")
-	showMCPPrompts := flags.Bool("mcp-prompts", false, "list daemon MCP prompts and exit without starting a run")
-	mcpPromptsJSON := flags.Bool("mcp-prompts-json", false, "print --mcp-prompts output as JSON")
-	showMCPRead := flags.Bool("mcp-read", false, "read one daemon MCP resource and exit without starting a run")
-	mcpReadJSON := flags.Bool("mcp-read-json", false, "print --mcp-read output as JSON")
-	showMCPPrompt := flags.Bool("mcp-prompt", false, "render one daemon MCP prompt and exit without starting a run")
-	mcpPromptJSON := flags.Bool("mcp-prompt-json", false, "print --mcp-prompt output as JSON")
-	mcpServer := flags.String("server", "", "MCP server name for --mcp-read or --mcp-prompt")
-	mcpURI := flags.String("uri", "", "MCP resource URI for --mcp-read")
-	mcpName := flags.String("name", "", "MCP prompt name for --mcp-prompt")
-	recallQuery := flags.String("recall", "", "search daemon recall history and exit without starting a run")
-	recallJSON := flags.Bool("recall-json", false, "print --recall output as JSON")
-	recallMemories := flags.Bool("recall-memories", false, "restrict --recall to curated memories")
-	recallLimit := flags.Int("recall-limit", 0, "maximum recall hits; 0 uses daemon default")
-	showMemories := flags.Bool("memories", false, "list daemon memories and exit without starting a run")
-	memoriesJSON := flags.Bool("memories-json", false, "print --memories output as JSON")
-	memoryLimit := flags.Int("memory-limit", 0, "maximum memories to return; 0 means no limit")
-	forgetMemoryID := flags.String("forget-memory", "", "delete one daemon memory by id and exit without starting a run")
-	forgetMemoryJSON := flags.Bool("forget-memory-json", false, "print --forget-memory output as JSON")
-	showPermissions := flags.Bool("permissions", false, "list daemon permission grants and exit without starting a run")
-	permissionsJSON := flags.Bool("permissions-json", false, "print --permissions output as JSON")
-	forgetPermissionID := flags.String("forget-permission", "", "delete one daemon permission grant by id and exit without starting a run")
-	forgetPermissionJSON := flags.Bool("forget-permission-json", false, "print --forget-permission output as JSON")
-	showStatus := flags.Bool("status", false, "show daemon status and exit without starting a run")
-	statusJSON := flags.Bool("status-json", false, "print --status output as JSON")
-	showDiagnose := flags.Bool("diagnose", false, "diagnose local daemon connectivity and runtime state without starting a run")
-	diagnoseJSON := flags.Bool("diagnose-json", false, "print --diagnose output as JSON")
-	showInspect := flags.Bool("inspect", false, "show daemon status and tools and exit without starting a run")
-	inspectJSON := flags.Bool("inspect-json", false, "print --inspect output as JSON")
-	var envs envFiles
-	flags.Var(&envs, "env", "env file path; repeatable")
-	var runArgs repeatedStrings
-	flags.Var(&runArgs, "arg", "skill or MCP prompt argument key=value; repeatable")
-
-	if err := flags.Parse(args); err != nil {
-		return err
-	}
-	markUsagePricingFlagState(flags, usagePricing)
-	if err := validateUsagePricing(*usagePricing); err != nil {
-		return err
-	}
-	if *toolsJSON {
-		*showTools = true
-	}
-	if *skillsJSON {
-		*showSkills = true
-	}
-	if *rolesJSON {
-		*showRoles = true
-	}
-	if *mcpResourcesJSON {
-		*showMCPResources = true
-	}
-	if *mcpPromptsJSON {
-		*showMCPPrompts = true
-	}
-	if *mcpReadJSON {
-		*showMCPRead = true
-	}
-	if *mcpPromptJSON {
-		*showMCPPrompt = true
-	}
-	if *statusJSON {
-		*showStatus = true
-	}
-	if *diagnoseJSON {
-		*showDiagnose = true
-	}
-	if *inspectJSON {
-		*showInspect = true
-	}
-	if *recallJSON && strings.TrimSpace(*recallQuery) == "" && flags.NArg() == 1 {
-		*recallQuery = flags.Arg(0)
-	}
-	if *memoriesJSON {
-		*showMemories = true
-	}
-	if *permissionsJSON {
-		*showPermissions = true
-	}
-	showRecall := strings.TrimSpace(*recallQuery) != "" || *recallJSON
-	showForgetMemory := strings.TrimSpace(*forgetMemoryID) != "" || *forgetMemoryJSON
-	showForgetPermission := strings.TrimSpace(*forgetPermissionID) != "" || *forgetPermissionJSON
-	inspectModes := 0
-	for _, enabled := range []bool{*showTools, *showSkills, *showRoles, *showMCPResources, *showMCPPrompts, *showMCPRead, *showMCPPrompt, showRecall, *showMemories, showForgetMemory, *showPermissions, showForgetPermission, *showStatus, *showDiagnose, *showInspect} {
-		if enabled {
-			inspectModes++
-		}
-	}
-	if inspectModes > 1 {
-		return errors.New("choose only one of --tools, --skills, --roles, --mcp-resources, --mcp-prompts, --mcp-read, --mcp-prompt, --recall, --memories, --forget-memory, --permissions, --forget-permission, --status, --diagnose, or --inspect")
-	}
-	if inspectModes == 0 && strings.TrimSpace(*prompt) == "" && strings.TrimSpace(*skillName) == "" {
-		return errors.New("missing required --prompt or --skill")
-	}
-	if inspectModes == 0 && strings.TrimSpace(*prompt) != "" && strings.TrimSpace(*skillName) != "" {
-		return errors.New("choose only one of --prompt or --skill")
-	}
-	if err := loadEnvFiles(envs); err != nil {
-		return err
-	}
-	runArgsMap, err := parseConnectArgs(runArgs)
-	if err != nil {
-		return err
-	}
-	rawCfg := connectConfig{
-		BaseURL:      *baseURL,
-		Token:        *tokenFlag,
-		MetadataPath: *metadataPath,
-		SessionID:    *sessionID,
-		Prompt:       *prompt,
-		Skill:        *skillName,
-		SkillArgs:    runArgsMap,
-		ClientID:     *clientID,
-		Model:        *model,
-		Role:         *role,
-		MaxTurns:     *maxTurns,
-	}
-	if *showDiagnose {
-		return runConnectDiagnose(ctx, rawCfg, *diagnoseJSON, stdout, client)
-	}
-	cfg, err := resolveConnectConfig(rawCfg)
-	if err != nil {
-		return err
-	}
-	if *showTools {
-		return runConnectTools(ctx, cfg, *toolsJSON, stdout, client)
-	}
-	if *showSkills {
-		return runConnectSkills(ctx, cfg, *skillsJSON, stdout, client)
-	}
-	if *showRoles {
-		return runConnectRoles(ctx, cfg, *rolesJSON, stdout, client)
-	}
-	if *showMCPResources {
-		return runConnectMCPResources(ctx, cfg, *mcpResourcesJSON, stdout, client)
-	}
-	if *showMCPPrompts {
-		return runConnectMCPPrompts(ctx, cfg, *mcpPromptsJSON, stdout, client)
-	}
-	if *showMCPRead {
-		return runConnectMCPRead(ctx, cfg, *mcpServer, *mcpURI, *mcpReadJSON, stdout, client)
-	}
-	if *showMCPPrompt {
-		return runConnectMCPPrompt(ctx, cfg, *mcpServer, *mcpName, runArgsMap, *mcpPromptJSON, stdout, client)
-	}
-	if showRecall {
-		return runConnectRecall(ctx, cfg, *recallQuery, *recallMemories, *recallLimit, *recallJSON, stdout, client)
-	}
-	if *showMemories {
-		return runConnectMemories(ctx, cfg, *memoryLimit, *memoriesJSON, stdout, client)
-	}
-	if showForgetMemory {
-		return runConnectForgetMemory(ctx, cfg, *forgetMemoryID, *forgetMemoryJSON, stdout, client)
-	}
-	if *showPermissions {
-		return runConnectPermissions(ctx, cfg, *permissionsJSON, stdout, client)
-	}
-	if showForgetPermission {
-		return runConnectForgetPermission(ctx, cfg, *forgetPermissionID, *forgetPermissionJSON, stdout, client)
-	}
-	if *showStatus {
-		return runConnectStatus(ctx, cfg, *statusJSON, stdout, client)
-	}
-	if *showInspect {
-		return runConnectInspect(ctx, cfg, *memoryLimit, *inspectJSON, stdout, client)
-	}
-	return runConnect(ctx, cfg, *showUsage, *usagePricing, stdin, stdout, stderr, client)
-}
-
-func resolveConnectConfig(cfg connectConfig) (connectConfig, error) {
-	var meta daemonMetadata
-	var metadataErr error
-	if strings.TrimSpace(cfg.MetadataPath) != "" {
-		loaded, err := readDaemonMetadata(cfg.MetadataPath)
-		if err != nil {
-			metadataErr = err
-		} else {
-			meta = loaded
-		}
-	}
-	if strings.TrimSpace(cfg.BaseURL) == "" {
-		cfg.BaseURL = meta.BaseURL
-	}
-	if strings.TrimSpace(cfg.Token) == "" {
-		cfg.Token = meta.Token
-	}
-	if strings.TrimSpace(cfg.Token) == "" {
-		cfg.Token = strings.TrimSpace(os.Getenv("GLUE_DAEMON_TOKEN"))
-	}
-	cfg.BaseURL = strings.TrimRight(strings.TrimSpace(cfg.BaseURL), "/")
-	cfg.Token = strings.TrimSpace(cfg.Token)
-	cfg.SessionID = strings.TrimSpace(cfg.SessionID)
-	cfg.ClientID = strings.TrimSpace(cfg.ClientID)
-	cfg.Prompt = strings.TrimSpace(cfg.Prompt)
-	cfg.Skill = strings.TrimSpace(cfg.Skill)
-	if cfg.SessionID == "" {
-		cfg.SessionID = "default"
-	}
-	if cfg.ClientID == "" {
-		cfg.ClientID = defaultClientID()
-	}
-	if cfg.BaseURL == "" {
-		if metadataErr != nil {
-			return cfg, metadataErr
-		}
-		return cfg, errors.New("daemon base URL is required (metadata missing or pass --base-url)")
-	}
-	if cfg.Token == "" {
-		if metadataErr != nil {
-			return cfg, metadataErr
-		}
-		return cfg, errors.New("daemon token is required (metadata missing, pass --token, or set GLUE_DAEMON_TOKEN)")
-	}
-	return cfg, nil
-}
-
-func runConnect(ctx context.Context, cfg connectConfig, showUsage bool, pricing usagePricing, stdin io.Reader, stdout io.Writer, stderr io.Writer, client httpDoer) error {
-	start, err := startDaemonRun(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	done := make(chan struct{})
-	defer close(done)
-	go func() {
-		select {
-		case <-ctx.Done():
-			_ = cancelDaemonRun(context.Background(), cfg, start.RunID, client)
-		case <-done:
-		}
-	}()
-	return streamDaemonRun(ctx, cfg, start, showUsage, pricing, bufio.NewReader(stdin), stdout, stderr, client)
-}
-
-func runConnectTools(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonTools(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonToolCatalog(stdout, catalog.Tools)
-	return nil
-}
-
-func runConnectSkills(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonSkills(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonSkillCatalog(stdout, catalog.Skills)
-	return nil
-}
-
-func runConnectRoles(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonRoles(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonRoleCatalog(stdout, catalog.Roles)
-	return nil
-}
-
-func runConnectMCPResources(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonMCPResources(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonMCPResourceCatalog(stdout, catalog.Resources)
-	return nil
-}
-
-func runConnectMCPPrompts(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonMCPPrompts(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonMCPPromptCatalog(stdout, catalog.Prompts)
-	return nil
-}
-
-func runConnectMCPRead(ctx context.Context, cfg connectConfig, server, uri string, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	server = strings.TrimSpace(server)
-	uri = strings.TrimSpace(uri)
-	if server == "" {
-		return errors.New("--server is required for --mcp-read")
-	}
-	if uri == "" {
-		return errors.New("--uri is required for --mcp-read")
-	}
-	read, err := requestDaemonMCPResourceRead(ctx, cfg, server, uri, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(read)
-	}
-	writeDaemonMCPResourceRead(stdout, read)
-	return nil
-}
-
-func runConnectMCPPrompt(ctx context.Context, cfg connectConfig, server, name string, args map[string]string, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	server = strings.TrimSpace(server)
-	name = strings.TrimSpace(name)
-	if server == "" {
-		return errors.New("--server is required for --mcp-prompt")
-	}
-	if name == "" {
-		return errors.New("--name is required for --mcp-prompt")
-	}
-	rendered, err := requestDaemonMCPPrompt(ctx, cfg, server, name, args, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(rendered)
-	}
-	writeDaemonMCPPrompt(stdout, rendered)
-	return nil
-}
-
-func runConnectRecall(ctx context.Context, cfg connectConfig, query string, memoriesOnly bool, limit int, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	query = strings.TrimSpace(query)
-	if query == "" {
-		return errors.New("--recall query is required")
-	}
-	if limit < 0 {
-		return errors.New("--recall-limit must be non-negative")
-	}
-	recall, err := requestDaemonRecall(ctx, cfg, daemon.RecallRequest{
-		Query:        query,
-		Limit:        limit,
-		MemoriesOnly: memoriesOnly,
-	}, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(recall)
-	}
-	writeDaemonRecallHits(stdout, recall.Hits)
-	return nil
-}
-
-func runConnectMemories(ctx context.Context, cfg connectConfig, limit int, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	if limit < 0 {
-		return errors.New("--memory-limit must be non-negative")
-	}
-	catalog, err := fetchDaemonMemories(ctx, cfg, limit, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonMemories(stdout, catalog.Memories)
-	return nil
-}
-
-func runConnectForgetMemory(ctx context.Context, cfg connectConfig, id string, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	id = strings.TrimSpace(id)
-	if id == "" {
-		return errors.New("--forget-memory is required")
-	}
-	forgotten, err := requestDaemonForgetMemory(ctx, cfg, id, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(forgotten)
-	}
-	writeDaemonMemory(stdout, forgotten.Memory)
-	return nil
-}
-
-func runConnectPermissions(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	catalog, err := fetchDaemonPermissions(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(catalog)
-	}
-	writeDaemonPermissions(stdout, catalog.Permissions)
-	return nil
-}
-
-func runConnectForgetPermission(ctx context.Context, cfg connectConfig, id string, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	id = strings.TrimSpace(id)
-	if id == "" {
-		return errors.New("--forget-permission is required")
-	}
-	forgotten, err := requestDaemonForgetPermission(ctx, cfg, id, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(forgotten)
-	}
-	writeDaemonPermission(stdout, forgotten.Permission)
-	return nil
-}
-
-func runConnectStatus(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	status, err := fetchDaemonStatus(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(status)
-	}
-	writeDaemonStatus(stdout, status)
-	return nil
-}
-
-func runConnectDiagnose(ctx context.Context, cfg connectConfig, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	diagnosis := diagnoseDaemon(ctx, cfg, client)
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(diagnosis)
-	}
-	writeDaemonDiagnosis(stdout, diagnosis)
-	return nil
-}
-
-func runConnectInspect(ctx context.Context, cfg connectConfig, memoryLimit int, jsonOutput bool, stdout io.Writer, client httpDoer) error {
-	if memoryLimit < 0 {
-		return errors.New("--memory-limit must be non-negative")
-	}
-	status, err := fetchDaemonStatus(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	catalog, err := fetchDaemonTools(ctx, cfg, client)
-	if err != nil {
-		return err
-	}
-	inspect := daemonInspect{Status: status, Tools: catalog.Tools}
-	if daemonHasCapability(status, "skills") {
-		skills, err := fetchDaemonSkills(ctx, cfg, client)
-		if err != nil {
-			return err
-		}
-		inspect.Skills = skills.Skills
-	}
-	if daemonHasCapability(status, "roles") {
-		roles, err := fetchDaemonRoles(ctx, cfg, client)
-		if err != nil {
-			return err
-		}
-		inspect.Roles = roles.Roles
-	}
-	if daemonHasCapability(status, "memories") {
-		memories, err := fetchDaemonMemories(ctx, cfg, memoryLimit, client)
-		if err != nil {
-			return err
-		}
-		inspect.Memories = memories.Memories
-	}
-	if daemonHasCapability(status, "permission_grants") {
-		permissions, err := fetchDaemonPermissions(ctx, cfg, client)
-		if err != nil {
-			return err
-		}
-		inspect.Permissions = permissions.Permissions
-	}
-	if daemonHasCapability(status, "mcp_resources") {
-		resources, err := fetchDaemonMCPResources(ctx, cfg, client)
-		if err != nil {
-			return err
-		}
-		inspect.MCPResources = resources.Resources
-	}
-	if daemonHasCapability(status, "mcp_prompts") {
-		prompts, err := fetchDaemonMCPPrompts(ctx, cfg, client)
-		if err != nil {
-			return err
-		}
-		inspect.MCPPrompts = prompts.Prompts
-	}
-	if jsonOutput {
-		enc := json.NewEncoder(stdout)
-		enc.SetIndent("", "  ")
-		return enc.Encode(inspect)
-	}
-	writeDaemonInspect(stdout, inspect)
-	return nil
-}
-
-func fetchDaemonStatus(ctx context.Context, cfg connectConfig, client httpDoer) (daemonStatus, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/status", nil)
-	if err != nil {
-		return daemonStatus{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonStatus{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonStatus{}, fmt.Errorf("daemon status: %s", httpStatusError(resp))
-	}
-	var status daemonStatus
-	if err := json.NewDecoder(resp.Body).Decode(&status); err != nil {
-		return daemonStatus{}, err
-	}
-	return status, nil
-}
-
-func diagnoseDaemon(ctx context.Context, cfg connectConfig, client httpDoer) daemonDiagnosis {
-	diagnosis := daemonDiagnosis{
-		State:        "unknown",
-		MetadataPath: strings.TrimSpace(cfg.MetadataPath),
-	}
-	var meta daemonMetadata
-	var metadataErr error
-	if diagnosis.MetadataPath != "" {
-		loaded, err := readDaemonMetadata(diagnosis.MetadataPath)
-		if err != nil {
-			metadataErr = err
-			diagnosis.MetadataError = err.Error()
-		} else {
-			meta = loaded
-			diagnosis.MetadataFound = true
-			diagnosis.MetadataPID = meta.PID
-		}
-	}
-
-	baseURL := strings.TrimRight(strings.TrimSpace(cfg.BaseURL), "/")
-	switch {
-	case baseURL != "":
-		diagnosis.BaseURLSource = "flag"
-	case strings.TrimSpace(meta.BaseURL) != "":
-		baseURL = strings.TrimRight(strings.TrimSpace(meta.BaseURL), "/")
-		diagnosis.BaseURLSource = "metadata"
-	}
-	diagnosis.BaseURL = baseURL
-
-	token := strings.TrimSpace(cfg.Token)
-	switch {
-	case token != "":
-		diagnosis.TokenSource = "flag"
-	case strings.TrimSpace(meta.Token) != "":
-		token = strings.TrimSpace(meta.Token)
-		diagnosis.TokenSource = "metadata"
-	case strings.TrimSpace(os.Getenv("GLUE_DAEMON_TOKEN")) != "":
-		token = strings.TrimSpace(os.Getenv("GLUE_DAEMON_TOKEN"))
-		diagnosis.TokenSource = "GLUE_DAEMON_TOKEN"
-	default:
-		diagnosis.TokenSource = "missing"
-	}
-
-	if baseURL == "" {
-		if metadataErr != nil {
-			diagnosis.State = "no_metadata"
-			diagnosis.Summary = "daemon metadata is missing or unreadable"
-			diagnosis.Suggestions = []string{"Start Peggy with `peggy serve` or pass --base-url and --token explicitly."}
-			return diagnosis
-		}
-		diagnosis.State = "missing_base_url"
-		diagnosis.Summary = "daemon base URL is not configured"
-		diagnosis.Suggestions = []string{"Start Peggy with `peggy serve` or pass --base-url."}
-		return diagnosis
-	}
-	if token == "" {
-		diagnosis.State = "missing_token"
-		diagnosis.Summary = "daemon bearer token is not configured"
-		diagnosis.Suggestions = []string{"Use daemon metadata, pass --token, or set GLUE_DAEMON_TOKEN."}
-		return diagnosis
-	}
-
-	status, diagnostics, httpStatus, err := requestDaemonDiagnostics(ctx, baseURL, token, client)
-	diagnosis.HTTPStatus = httpStatus
-	if err != nil {
-		diagnosis.Summary = err.Error()
-		switch {
-		case httpStatus == http.StatusUnauthorized:
-			diagnosis.State = "auth_failed"
-			diagnosis.Suggestions = []string{"Use the token from the active metadata file or restart the daemon to refresh metadata."}
-		case httpStatus > 0:
-			diagnosis.State = "unhealthy"
-			diagnosis.Suggestions = []string{"Check daemon logs and restart `peggy serve` if the error persists."}
-		case diagnosis.MetadataFound && strings.TrimSpace(cfg.BaseURL) == "":
-			diagnosis.State = "stale_metadata"
-			diagnosis.Suggestions = []string{"The metadata file points at an unreachable daemon; restart `peggy serve` to refresh it."}
-		default:
-			diagnosis.State = "unreachable"
-			diagnosis.Suggestions = []string{"Check that the daemon is running and the --base-url value is correct."}
-		}
-		return diagnosis
-	}
-	diagnosis.OK = true
-	diagnosis.State = "healthy"
-	diagnosis.Summary = "daemon is reachable and authenticated"
-	diagnosis.Status = &status
-	diagnosis.Diagnostics = &diagnostics
-	return diagnosis
-}
-
-func requestDaemonDiagnostics(ctx context.Context, baseURL, token string, client httpDoer) (daemonStatus, daemonDiagnostics, int, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, baseURL+"/v1/diagnostics", nil)
-	if err != nil {
-		return daemonStatus{}, daemonDiagnostics{}, 0, err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonStatus{}, daemonDiagnostics{}, 0, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonStatus{}, daemonDiagnostics{}, resp.StatusCode, fmt.Errorf("daemon diagnostics: %s", httpStatusError(resp))
-	}
-	var diagnostics daemonDiagnostics
-	if err := json.NewDecoder(resp.Body).Decode(&diagnostics); err != nil {
-		return daemonStatus{}, daemonDiagnostics{}, resp.StatusCode, err
-	}
-	status := daemonStatus{
-		OK:           diagnostics.OK,
-		Version:      diagnostics.Version,
-		ActiveRuns:   diagnostics.ActiveRuns,
-		ToolsCount:   diagnostics.ToolsCount,
-		Capabilities: diagnostics.Capabilities,
-	}
-	return status, diagnostics, resp.StatusCode, nil
-}
-
-func writeDaemonStatus(w io.Writer, status daemonStatus) {
-	state := "error"
-	if status.OK {
-		state = "ok"
-	}
-	fmt.Fprintf(w, "status: %s\n", state)
-	fmt.Fprintf(w, "version: %d\n", status.Version)
-	fmt.Fprintf(w, "active_runs: %d\n", status.ActiveRuns)
-	fmt.Fprintf(w, "tools_count: %d\n", status.ToolsCount)
-	if len(status.Capabilities) > 0 {
-		fmt.Fprintf(w, "capabilities: %s\n", strings.Join(status.Capabilities, ", "))
-	}
-}
-
-func writeDaemonDiagnosis(w io.Writer, diagnosis daemonDiagnosis) {
-	state := diagnosis.State
-	if state == "" {
-		state = "unknown"
-	}
-	fmt.Fprintf(w, "daemon: %s\n", state)
-	if diagnosis.Summary != "" {
-		fmt.Fprintf(w, "summary: %s\n", diagnosis.Summary)
-	}
-	if diagnosis.MetadataPath != "" {
-		fmt.Fprintf(w, "metadata: %s", diagnosis.MetadataPath)
-		if diagnosis.MetadataFound {
-			fmt.Fprint(w, " (found)")
-			if diagnosis.MetadataPID != 0 {
-				fmt.Fprintf(w, " pid=%d", diagnosis.MetadataPID)
-			}
-		} else {
-			fmt.Fprint(w, " (missing)")
-		}
-		fmt.Fprintln(w)
-	}
-	if diagnosis.MetadataError != "" {
-		fmt.Fprintf(w, "metadata_error: %s\n", diagnosis.MetadataError)
-	}
-	if diagnosis.BaseURL != "" {
-		fmt.Fprintf(w, "base_url: %s", diagnosis.BaseURL)
-		if diagnosis.BaseURLSource != "" {
-			fmt.Fprintf(w, " (%s)", diagnosis.BaseURLSource)
-		}
-		fmt.Fprintln(w)
-	}
-	if diagnosis.TokenSource != "" {
-		fmt.Fprintf(w, "token: %s\n", diagnosis.TokenSource)
-	}
-	if diagnosis.HTTPStatus != 0 {
-		fmt.Fprintf(w, "http_status: %d\n", diagnosis.HTTPStatus)
-	}
-	if diagnosis.Status != nil {
-		fmt.Fprintf(w, "active_runs: %d\n", diagnosis.Status.ActiveRuns)
-		fmt.Fprintf(w, "tools_count: %d\n", diagnosis.Status.ToolsCount)
-	}
-	if diagnosis.Diagnostics != nil {
-		runtime := diagnosis.Diagnostics.Runtime
-		if runtime.Provider != "" {
-			fmt.Fprintf(w, "provider: %s\n", runtime.Provider)
-		}
-		if runtime.Model != "" {
-			fmt.Fprintf(w, "model: %s\n", runtime.Model)
-		}
-		if runtime.StoreType != "" || runtime.StorePath != "" {
-			fmt.Fprintf(w, "store: %s", runtime.StoreType)
-			if runtime.StorePath != "" {
-				fmt.Fprintf(w, " %s", runtime.StorePath)
-			}
-			fmt.Fprintln(w)
-		}
-		if runtime.ListenAddr != "" {
-			fmt.Fprintf(w, "listen_addr: %s\n", runtime.ListenAddr)
-		}
-		if runtime.MetadataPath != "" {
-			fmt.Fprintf(w, "runtime_metadata: %s\n", runtime.MetadataPath)
-		}
-		if runtime.TokenSource != "" {
-			fmt.Fprintf(w, "runtime_token_source: %s\n", runtime.TokenSource)
-		}
-		if runtime.CodingEnabled {
-			fmt.Fprintf(w, "coding: enabled")
-			if runtime.CodingWorkDir != "" {
-				fmt.Fprintf(w, " %s", runtime.CodingWorkDir)
-			}
-			fmt.Fprintln(w)
-		}
-		if len(diagnosis.Diagnostics.RecentErrors) > 0 {
-			fmt.Fprintln(w, "recent_errors:")
-			for _, entry := range diagnosis.Diagnostics.RecentErrors {
-				timestamp := "unknown"
-				if !entry.Time.IsZero() {
-					timestamp = entry.Time.Format(time.RFC3339)
-				}
-				fmt.Fprintf(w, "  %s %s", timestamp, entry.Error)
-				if entry.RunID != "" {
-					fmt.Fprintf(w, " run=%s", entry.RunID)
-				}
-				if entry.SessionID != "" {
-					fmt.Fprintf(w, " session=%s", entry.SessionID)
-				}
-				if entry.ClientID != "" {
-					fmt.Fprintf(w, " client=%s", entry.ClientID)
-				}
-				fmt.Fprintln(w)
-			}
-		}
-	}
-	if len(diagnosis.Suggestions) > 0 {
-		fmt.Fprintln(w, "suggestions:")
-		for _, suggestion := range diagnosis.Suggestions {
-			fmt.Fprintf(w, "  - %s\n", suggestion)
-		}
-	}
-}
-
-func writeDaemonInspect(w io.Writer, inspect daemonInspect) {
-	writeDaemonStatus(w, inspect.Status)
-	fmt.Fprintln(w)
-	fmt.Fprintln(w, "tools:")
-	writeDaemonToolCatalogIndented(w, inspect.Tools, "  ")
-	if daemonHasCapability(inspect.Status, "skills") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "skills:")
-		writeDaemonSkillCatalogIndented(w, inspect.Skills, "  ")
-	}
-	if daemonHasCapability(inspect.Status, "roles") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "roles:")
-		writeDaemonRoleCatalogIndented(w, inspect.Roles, "  ")
-	}
-	if daemonHasCapability(inspect.Status, "memories") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "memories:")
-		writeDaemonMemoriesIndented(w, inspect.Memories, "  ")
-	}
-	if daemonHasCapability(inspect.Status, "permission_grants") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "permissions:")
-		writeDaemonPermissionsIndented(w, inspect.Permissions, "  ")
-	}
-	if daemonHasCapability(inspect.Status, "mcp_resources") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "mcp_resources:")
-		writeDaemonMCPResourceCatalogIndented(w, inspect.MCPResources, "  ")
-	}
-	if daemonHasCapability(inspect.Status, "mcp_prompts") {
-		fmt.Fprintln(w)
-		fmt.Fprintln(w, "mcp_prompts:")
-		writeDaemonMCPPromptCatalogIndented(w, inspect.MCPPrompts, "  ")
-	}
-}
-
-func fetchDaemonTools(ctx context.Context, cfg connectConfig, client httpDoer) (daemonToolCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/tools", nil)
-	if err != nil {
-		return daemonToolCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonToolCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonToolCatalog{}, fmt.Errorf("daemon tools: %s", httpStatusError(resp))
-	}
-	var catalog daemonToolCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonToolCatalog{}, err
-	}
-	if catalog.Tools == nil {
-		catalog.Tools = []daemonToolCatalogEntry{}
-	}
-	return catalog, nil
-}
-
-func fetchDaemonSkills(ctx context.Context, cfg connectConfig, client httpDoer) (daemonSkillCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/skills", nil)
-	if err != nil {
-		return daemonSkillCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonSkillCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonSkillCatalog{}, fmt.Errorf("daemon skills: %s", httpStatusError(resp))
-	}
-	var catalog daemonSkillCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonSkillCatalog{}, err
-	}
-	if catalog.Skills == nil {
-		catalog.Skills = []daemon.SkillCatalogEntry{}
-	}
-	return catalog, nil
-}
-
-func fetchDaemonRoles(ctx context.Context, cfg connectConfig, client httpDoer) (daemonRoleCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/roles", nil)
-	if err != nil {
-		return daemonRoleCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonRoleCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonRoleCatalog{}, fmt.Errorf("daemon roles: %s", httpStatusError(resp))
-	}
-	var catalog daemonRoleCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonRoleCatalog{}, err
-	}
-	if catalog.Roles == nil {
-		catalog.Roles = []daemon.RoleCatalogEntry{}
-	}
-	return catalog, nil
-}
-
-func fetchDaemonMemories(ctx context.Context, cfg connectConfig, limit int, client httpDoer) (daemon.MemoryCatalogResponse, error) {
-	endpoint := cfg.BaseURL + "/v1/memories"
-	if limit > 0 {
-		values := url.Values{}
-		values.Set("limit", fmt.Sprint(limit))
-		endpoint += "?" + values.Encode()
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
-	if err != nil {
-		return daemon.MemoryCatalogResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.MemoryCatalogResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.MemoryCatalogResponse{}, fmt.Errorf("daemon memories: %s", httpStatusError(resp))
-	}
-	var catalog daemon.MemoryCatalogResponse
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemon.MemoryCatalogResponse{}, err
-	}
-	if catalog.Memories == nil {
-		catalog.Memories = []daemon.MemoryEntry{}
-	}
-	return catalog, nil
-}
-
-func requestDaemonForgetMemory(ctx context.Context, cfg connectConfig, id string, client httpDoer) (daemon.MemoryForgetResponse, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodDelete, cfg.BaseURL+"/v1/memories/"+url.PathEscape(id), nil)
-	if err != nil {
-		return daemon.MemoryForgetResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.MemoryForgetResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.MemoryForgetResponse{}, fmt.Errorf("daemon memory delete: %s", httpStatusError(resp))
-	}
-	var forgotten daemon.MemoryForgetResponse
-	if err := json.NewDecoder(resp.Body).Decode(&forgotten); err != nil {
-		return daemon.MemoryForgetResponse{}, err
-	}
-	return forgotten, nil
-}
-
-func fetchDaemonPermissions(ctx context.Context, cfg connectConfig, client httpDoer) (daemonPermissionCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/permissions", nil)
-	if err != nil {
-		return daemonPermissionCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonPermissionCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonPermissionCatalog{}, fmt.Errorf("daemon permissions: %s", httpStatusError(resp))
-	}
-	var catalog daemonPermissionCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonPermissionCatalog{}, err
-	}
-	if catalog.Permissions == nil {
-		catalog.Permissions = []daemon.PermissionGrant{}
-	}
-	return catalog, nil
-}
-
-func requestDaemonForgetPermission(ctx context.Context, cfg connectConfig, id string, client httpDoer) (daemon.PermissionForgetResponse, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodDelete, cfg.BaseURL+"/v1/permissions/"+url.PathEscape(id), nil)
-	if err != nil {
-		return daemon.PermissionForgetResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.PermissionForgetResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.PermissionForgetResponse{}, fmt.Errorf("daemon permission delete: %s", httpStatusError(resp))
-	}
-	var forgotten daemon.PermissionForgetResponse
-	if err := json.NewDecoder(resp.Body).Decode(&forgotten); err != nil {
-		return daemon.PermissionForgetResponse{}, err
-	}
-	return forgotten, nil
-}
-
-func fetchDaemonMCPResources(ctx context.Context, cfg connectConfig, client httpDoer) (daemonMCPResourceCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/mcp/resources", nil)
-	if err != nil {
-		return daemonMCPResourceCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonMCPResourceCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonMCPResourceCatalog{}, fmt.Errorf("daemon MCP resources: %s", httpStatusError(resp))
-	}
-	var catalog daemonMCPResourceCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonMCPResourceCatalog{}, err
-	}
-	if catalog.Resources == nil {
-		catalog.Resources = []daemon.MCPResourceCatalogEntry{}
-	}
-	return catalog, nil
-}
-
-func fetchDaemonMCPPrompts(ctx context.Context, cfg connectConfig, client httpDoer) (daemonMCPPromptCatalog, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+"/v1/mcp/prompts", nil)
-	if err != nil {
-		return daemonMCPPromptCatalog{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemonMCPPromptCatalog{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemonMCPPromptCatalog{}, fmt.Errorf("daemon MCP prompts: %s", httpStatusError(resp))
-	}
-	var catalog daemonMCPPromptCatalog
-	if err := json.NewDecoder(resp.Body).Decode(&catalog); err != nil {
-		return daemonMCPPromptCatalog{}, err
-	}
-	if catalog.Prompts == nil {
-		catalog.Prompts = []daemon.MCPPromptCatalogEntry{}
-	}
-	return catalog, nil
-}
-
-func requestDaemonMCPResourceRead(ctx context.Context, cfg connectConfig, server, uri string, client httpDoer) (daemon.MCPResourceReadResponse, error) {
-	payload := daemon.MCPReadResourceRequest{Server: server, URI: uri}
-	var body bytes.Buffer
-	if err := json.NewEncoder(&body).Encode(payload); err != nil {
-		return daemon.MCPResourceReadResponse{}, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, cfg.BaseURL+"/v1/mcp/resources/read", &body)
-	if err != nil {
-		return daemon.MCPResourceReadResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.MCPResourceReadResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.MCPResourceReadResponse{}, fmt.Errorf("daemon MCP resource read: %s", httpStatusError(resp))
-	}
-	var read daemon.MCPResourceReadResponse
-	if err := json.NewDecoder(resp.Body).Decode(&read); err != nil {
-		return daemon.MCPResourceReadResponse{}, err
-	}
-	if read.Contents == nil {
-		read.Contents = []daemon.MCPResourceContent{}
-	}
-	return read, nil
-}
-
-func requestDaemonMCPPrompt(ctx context.Context, cfg connectConfig, server, name string, args map[string]string, client httpDoer) (daemon.MCPPromptRenderResponse, error) {
-	payload := daemon.MCPPromptRenderRequest{Server: server, Name: name, Arguments: args}
-	var body bytes.Buffer
-	if err := json.NewEncoder(&body).Encode(payload); err != nil {
-		return daemon.MCPPromptRenderResponse{}, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, cfg.BaseURL+"/v1/mcp/prompts/get", &body)
-	if err != nil {
-		return daemon.MCPPromptRenderResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.MCPPromptRenderResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.MCPPromptRenderResponse{}, fmt.Errorf("daemon MCP prompt: %s", httpStatusError(resp))
-	}
-	var rendered daemon.MCPPromptRenderResponse
-	if err := json.NewDecoder(resp.Body).Decode(&rendered); err != nil {
-		return daemon.MCPPromptRenderResponse{}, err
-	}
-	if rendered.Messages == nil {
-		rendered.Messages = []daemon.MCPPromptMessage{}
-	}
-	return rendered, nil
-}
-
-func requestDaemonRecall(ctx context.Context, cfg connectConfig, request daemon.RecallRequest, client httpDoer) (daemon.RecallResponse, error) {
-	var body bytes.Buffer
-	if err := json.NewEncoder(&body).Encode(request); err != nil {
-		return daemon.RecallResponse{}, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, cfg.BaseURL+"/v1/recall", &body)
-	if err != nil {
-		return daemon.RecallResponse{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return daemon.RecallResponse{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return daemon.RecallResponse{}, fmt.Errorf("daemon recall: %s", httpStatusError(resp))
-	}
-	var recall daemon.RecallResponse
-	if err := json.NewDecoder(resp.Body).Decode(&recall); err != nil {
-		return daemon.RecallResponse{}, err
-	}
-	if recall.Hits == nil {
-		recall.Hits = []daemon.RecallHit{}
-	}
-	return recall, nil
-}
-
-func writeDaemonToolCatalog(w io.Writer, tools []daemonToolCatalogEntry) {
-	writeDaemonToolCatalogIndented(w, tools, "")
-}
-
-func writeDaemonToolCatalogIndented(w io.Writer, tools []daemonToolCatalogEntry, indent string) {
-	if len(tools) == 0 {
-		fmt.Fprintf(w, "%sNo daemon tools reported.\n", indent)
-		return
-	}
-	for i, tool := range tools {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%s\n", indent, tool.Name)
-		if tool.Description != "" {
-			fmt.Fprintf(w, "%s  description: %s\n", indent, oneLine(tool.Description))
-		}
-		if tool.RequiresPermission || tool.PermissionAction != "" || tool.PermissionTargetPreview != "" {
-			fmt.Fprintf(w, "%s  permission: %s %s\n", indent, tool.PermissionAction, tool.PermissionTargetPreview)
-		}
-		if len(tool.Parameters) > 0 {
-			fmt.Fprintf(w, "%s  parameters: %s\n", indent, compactJSON(tool.Parameters))
-		}
-	}
-}
-
-func writeDaemonSkillCatalog(w io.Writer, skills []daemon.SkillCatalogEntry) {
-	writeDaemonSkillCatalogIndented(w, skills, "")
-}
-
-func writeDaemonSkillCatalogIndented(w io.Writer, skills []daemon.SkillCatalogEntry, indent string) {
-	if len(skills) == 0 {
-		fmt.Fprintf(w, "%sNo daemon skills reported.\n", indent)
-		return
-	}
-	for i, skill := range skills {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%s\n", indent, skill.Name)
-		if skill.Description != "" {
-			fmt.Fprintf(w, "%s  description: %s\n", indent, oneLine(skill.Description))
-		}
-	}
-}
-
-func writeDaemonRoleCatalog(w io.Writer, roles []daemon.RoleCatalogEntry) {
-	writeDaemonRoleCatalogIndented(w, roles, "")
-}
-
-func writeDaemonRoleCatalogIndented(w io.Writer, roles []daemon.RoleCatalogEntry, indent string) {
-	if len(roles) == 0 {
-		fmt.Fprintf(w, "%sNo daemon roles reported.\n", indent)
-		return
-	}
-	for i, role := range roles {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%s\n", indent, role.Name)
-		if role.Description != "" {
-			fmt.Fprintf(w, "%s  description: %s\n", indent, oneLine(role.Description))
-		}
-		if role.Model != "" {
-			fmt.Fprintf(w, "%s  model: %s\n", indent, role.Model)
-		}
-	}
-}
-
-func writeDaemonMemories(w io.Writer, memories []daemon.MemoryEntry) {
-	writeDaemonMemoriesIndented(w, memories, "")
-}
-
-func writeDaemonMemoriesIndented(w io.Writer, memories []daemon.MemoryEntry, indent string) {
-	if len(memories) == 0 {
-		fmt.Fprintf(w, "%sNo daemon memories reported.\n", indent)
-		return
-	}
-	for i, memory := range memories {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%d. %s\n", indent, i+1, memory.ID)
-		writeDaemonMemoryFields(w, memory, indent)
-	}
-}
-
-func writeDaemonMemory(w io.Writer, memory daemon.MemoryEntry) {
-	fmt.Fprintln(w, memory.ID)
-	writeDaemonMemoryFields(w, memory, "")
-}
-
-func writeDaemonMemoryFields(w io.Writer, memory daemon.MemoryEntry, indent string) {
-	if !memory.Timestamp.IsZero() {
-		fmt.Fprintf(w, "%s  timestamp: %s\n", indent, memory.Timestamp.Format(time.RFC3339))
-	}
-	fmt.Fprintf(w, "%s  content: %s\n", indent, oneLine(memory.Content))
-	if len(memory.Tags) > 0 {
-		fmt.Fprintf(w, "%s  tags: %s\n", indent, strings.Join(memory.Tags, ", "))
-	}
-}
-
-func writeDaemonPermissions(w io.Writer, permissions []daemon.PermissionGrant) {
-	writeDaemonPermissionsIndented(w, permissions, "")
-}
-
-func writeDaemonPermissionsIndented(w io.Writer, permissions []daemon.PermissionGrant, indent string) {
-	if len(permissions) == 0 {
-		fmt.Fprintf(w, "%sNo daemon permissions remembered.\n", indent)
-		return
-	}
-	for i, permission := range permissions {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%d. %s\n", indent, i+1, permission.ID)
-		writeDaemonPermissionFields(w, permission, indent)
-	}
-}
-
-func writeDaemonPermission(w io.Writer, permission daemon.PermissionGrant) {
-	fmt.Fprintln(w, permission.ID)
-	writeDaemonPermissionFields(w, permission, "")
-}
-
-func writeDaemonPermissionFields(w io.Writer, permission daemon.PermissionGrant, indent string) {
-	fmt.Fprintf(w, "%s  scope: %s\n", indent, permission.Scope)
-	if permission.Owner != "" {
-		fmt.Fprintf(w, "%s  owner: %s\n", indent, permission.Owner)
-	}
-	if permission.ClientID != "" {
-		fmt.Fprintf(w, "%s  client_id: %s\n", indent, permission.ClientID)
-	}
-	if permission.SessionID != "" {
-		fmt.Fprintf(w, "%s  session_id: %s\n", indent, permission.SessionID)
-	}
-	fmt.Fprintf(w, "%s  tool: %s\n", indent, permission.Tool)
-	fmt.Fprintf(w, "%s  action: %s\n", indent, permission.Action)
-	if permission.Target != "" {
-		fmt.Fprintf(w, "%s  target: %s\n", indent, permission.Target)
-	}
-	if !permission.CreatedAt.IsZero() {
-		fmt.Fprintf(w, "%s  created_at: %s\n", indent, permission.CreatedAt.Format(time.RFC3339))
-	}
-}
-
-func writeDaemonMCPResourceCatalog(w io.Writer, resources []daemon.MCPResourceCatalogEntry) {
-	writeDaemonMCPResourceCatalogIndented(w, resources, "")
-}
-
-func writeDaemonMCPResourceCatalogIndented(w io.Writer, resources []daemon.MCPResourceCatalogEntry, indent string) {
-	if len(resources) == 0 {
-		fmt.Fprintf(w, "%sNo daemon MCP resources reported.\n", indent)
-		return
-	}
-	for i, resource := range resources {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%s\n", indent, resource.URI)
-		fmt.Fprintf(w, "%s  server: %s\n", indent, resource.Server)
-		fmt.Fprintf(w, "%s  name: %s\n", indent, resource.Name)
-		if resource.Title != "" {
-			fmt.Fprintf(w, "%s  title: %s\n", indent, oneLine(resource.Title))
-		}
-		if resource.Description != "" {
-			fmt.Fprintf(w, "%s  description: %s\n", indent, oneLine(resource.Description))
-		}
-		if resource.MIMEType != "" {
-			fmt.Fprintf(w, "%s  mime_type: %s\n", indent, resource.MIMEType)
-		}
-		if resource.Size != nil {
-			fmt.Fprintf(w, "%s  size: %d\n", indent, *resource.Size)
-		}
-		if len(resource.Annotations) > 0 {
-			raw, err := json.Marshal(resource.Annotations)
-			if err == nil {
-				fmt.Fprintf(w, "%s  annotations: %s\n", indent, compactJSON(raw))
-			}
-		}
-	}
-}
-
-func writeDaemonMCPPromptCatalog(w io.Writer, prompts []daemon.MCPPromptCatalogEntry) {
-	writeDaemonMCPPromptCatalogIndented(w, prompts, "")
-}
-
-func writeDaemonMCPPromptCatalogIndented(w io.Writer, prompts []daemon.MCPPromptCatalogEntry, indent string) {
-	if len(prompts) == 0 {
-		fmt.Fprintf(w, "%sNo daemon MCP prompts reported.\n", indent)
-		return
-	}
-	for i, prompt := range prompts {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintf(w, "%s%s\n", indent, prompt.Name)
-		fmt.Fprintf(w, "%s  server: %s\n", indent, prompt.Server)
-		if prompt.Title != "" {
-			fmt.Fprintf(w, "%s  title: %s\n", indent, oneLine(prompt.Title))
-		}
-		if prompt.Description != "" {
-			fmt.Fprintf(w, "%s  description: %s\n", indent, oneLine(prompt.Description))
-		}
-		if len(prompt.Arguments) > 0 {
-			fmt.Fprintf(w, "%s  arguments:\n", indent)
-			for _, arg := range prompt.Arguments {
-				required := ""
-				if arg.Required {
-					required = " required"
-				}
-				line := arg.Name + required
-				if arg.Description != "" {
-					line += " - " + oneLine(arg.Description)
-				}
-				fmt.Fprintf(w, "%s    %s\n", indent, line)
-			}
-		}
-	}
-}
-
-func writeDaemonMCPResourceRead(w io.Writer, read daemon.MCPResourceReadResponse) {
-	if len(read.Contents) == 0 {
-		fmt.Fprintln(w, "No daemon MCP resource contents returned.")
-		return
-	}
-	for i, item := range read.Contents {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		fmt.Fprintln(w, item.URI)
-		fmt.Fprintf(w, "  server: %s\n", read.Server)
-		fmt.Fprintf(w, "  requested_uri: %s\n", read.URI)
-		if item.MIMEType != "" {
-			fmt.Fprintf(w, "  mime_type: %s\n", item.MIMEType)
-		}
-		if len(item.Meta) > 0 {
-			raw, err := json.Marshal(item.Meta)
-			if err == nil {
-				fmt.Fprintf(w, "  meta: %s\n", compactJSON(raw))
-			}
-		}
-		switch {
-		case item.Text != nil:
-			fmt.Fprintln(w, "  text:")
-			for _, line := range strings.Split(*item.Text, "\n") {
-				fmt.Fprintf(w, "    %s\n", line)
-			}
-		case item.Blob != nil:
-			fmt.Fprintf(w, "  blob: %s\n", *item.Blob)
-		}
-	}
-}
-
-func writeDaemonMCPPrompt(w io.Writer, rendered daemon.MCPPromptRenderResponse) {
-	fmt.Fprintln(w, rendered.Name)
-	fmt.Fprintf(w, "  server: %s\n", rendered.Server)
-	if rendered.Description != "" {
-		fmt.Fprintf(w, "  description: %s\n", oneLine(rendered.Description))
-	}
-	if len(rendered.Messages) == 0 {
-		fmt.Fprintln(w, "  messages: []")
-		return
-	}
-	fmt.Fprintln(w, "  messages:")
-	for _, message := range rendered.Messages {
-		fmt.Fprintf(w, "    - role: %s\n", message.Role)
-		if text, ok := daemonPromptTextContent(message.Content); ok {
-			fmt.Fprintln(w, "      text:")
-			for _, line := range strings.Split(text, "\n") {
-				fmt.Fprintf(w, "        %s\n", line)
-			}
-			continue
-		}
-		fmt.Fprintf(w, "      content: %s\n", compactJSON(message.Content))
-	}
-}
-
-func writeDaemonRecallHits(w io.Writer, hits []daemon.RecallHit) {
-	if len(hits) == 0 {
-		fmt.Fprintln(w, "No recall hits.")
-		return
-	}
-	for i, hit := range hits {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-		timestamp := ""
-		if !hit.Timestamp.IsZero() {
-			timestamp = hit.Timestamp.Format(time.RFC3339)
-		}
-		fmt.Fprintf(w, "%d. %s#%d\n", i+1, hit.SessionID, hit.Index)
-		if timestamp != "" {
-			fmt.Fprintf(w, "  timestamp: %s\n", timestamp)
-		}
-		fmt.Fprintf(w, "  score: %.2f\n", hit.Score)
-		fmt.Fprintf(w, "  snippet: %s\n", oneLine(hit.Snippet))
-	}
-}
-
-func daemonPromptTextContent(raw json.RawMessage) (string, bool) {
-	var content struct {
-		Type string `json:"type"`
-		Text string `json:"text"`
-	}
-	if err := json.Unmarshal(raw, &content); err != nil {
-		return "", false
-	}
-	if content.Type != "text" {
-		return "", false
-	}
-	return content.Text, true
-}
-
-func daemonHasCapability(status daemonStatus, capability string) bool {
-	for _, existing := range status.Capabilities {
-		if existing == capability {
-			return true
-		}
-	}
-	return false
-}
-
-func parseConnectArgs(values []string) (map[string]string, error) {
-	if len(values) == 0 {
-		return nil, nil
-	}
-	args := make(map[string]string, len(values))
-	for _, raw := range values {
-		key, value, ok := strings.Cut(raw, "=")
-		key = strings.TrimSpace(key)
-		if !ok || key == "" {
-			return nil, fmt.Errorf("--arg must be key=value")
-		}
-		args[key] = value
-	}
-	return args, nil
-}
-
-func oneLine(s string) string {
-	return strings.Join(strings.Fields(s), " ")
-}
-
-func compactJSON(raw json.RawMessage) string {
-	var buf bytes.Buffer
-	if err := json.Compact(&buf, raw); err != nil {
-		return oneLine(string(raw))
-	}
-	return buf.String()
-}
-
-func startDaemonRun(ctx context.Context, cfg connectConfig, client httpDoer) (startRunResult, error) {
-	payload := startRunPayload{
-		Text:      cfg.Prompt,
-		Skill:     cfg.Skill,
-		Arguments: cfg.SkillArgs,
-		ClientID:  cfg.ClientID,
-		Role:      strings.TrimSpace(cfg.Role),
-		Model:     strings.TrimSpace(cfg.Model),
-		MaxTurns:  cfg.MaxTurns,
-	}
-	var body bytes.Buffer
-	if err := json.NewEncoder(&body).Encode(payload); err != nil {
-		return startRunResult{}, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, cfg.BaseURL+"/v1/sessions/"+url.PathEscape(cfg.SessionID)+"/runs", &body)
-	if err != nil {
-		return startRunResult{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return startRunResult{}, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusCreated {
-		return startRunResult{}, fmt.Errorf("daemon start run: %s", httpStatusError(resp))
-	}
-	var out startRunResult
-	if err := json.NewDecoder(resp.Body).Decode(&out); err != nil {
-		return startRunResult{}, err
-	}
-	if out.RunID == "" || out.EventsURL == "" {
-		return startRunResult{}, errors.New("daemon start run: missing run id or events URL")
-	}
-	return out, nil
-}
-
-func streamDaemonRun(ctx context.Context, cfg connectConfig, start startRunResult, showUsage bool, pricing usagePricing, input *bufio.Reader, stdout io.Writer, stderr io.Writer, client httpDoer) error {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cfg.BaseURL+start.EventsURL, nil)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Accept", "text/event-stream")
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("daemon event stream: %s", httpStatusError(resp))
-	}
-
-	wroteDelta := false
-	scanner := bufio.NewScanner(resp.Body)
-	scanner.Buffer(make([]byte, 0, 64*1024), 1024*1024)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if !strings.HasPrefix(line, "data: ") {
-			continue
-		}
-		var event daemon.EventEnvelope
-		dec := json.NewDecoder(strings.NewReader(strings.TrimPrefix(line, "data: ")))
-		dec.UseNumber()
-		if err := dec.Decode(&event); err != nil {
-			return err
-		}
-		switch event.Type {
-		case string(glue.EventTextDelta):
-			if delta := payloadString(event.Payload, "delta"); delta != "" {
-				fmt.Fprint(stdout, delta)
-				wroteDelta = true
-			}
-		case "permission_request":
-			if wroteDelta {
-				fmt.Fprintln(stdout)
-				wroteDelta = false
-			}
-			perm, err := decodePayload[connectPermissionPayload](event.Payload)
-			if err != nil {
-				return err
-			}
-			decision, err := promptPermission(input, stderr, perm)
-			if err != nil {
-				return err
-			}
-			if err := postPermissionDecision(ctx, cfg, start.RunID, perm.PermissionID, decision, client); err != nil {
-				return err
-			}
-		case "run_done":
-			if wroteDelta {
-				fmt.Fprintln(stdout)
-			} else {
-				done, err := decodePayload[connectRunDonePayload](event.Payload)
-				if err != nil {
-					return err
-				}
-				if done.Text != "" {
-					fmt.Fprintln(stdout, done.Text)
-				}
-			}
-			if showUsage {
-				done, err := decodePayload[connectRunDonePayload](event.Payload)
-				if err != nil {
-					return err
-				}
-				messages := done.NewMessages
-				if len(messages) == 0 && done.Message != nil {
-					messages = []glue.Message{*done.Message}
-				}
-				writeUsageSummary(stderr, summarizeUsage(messages), pricing)
-			}
-			return nil
-		case "run_error":
-			if wroteDelta {
-				fmt.Fprintln(stdout)
-			}
-			if msg := payloadErrorMessage(event.Payload); msg != "" {
-				return errors.New(msg)
-			}
-			return errors.New("daemon run failed")
-		}
-	}
-	if err := scanner.Err(); err != nil {
-		return err
-	}
-	return errors.New("daemon event stream closed before terminal event")
-}
-
-func summarizeUsage(messages []glue.Message) usageSummary {
-	var summary usageSummary
-	for _, message := range messages {
-		if message.Role != glue.MessageRoleAssistant || message.Usage == nil {
-			continue
-		}
-		usage := message.Usage
-		summary.HasUsage = true
-		summary.InputTokens += usage.InputTokens
-		summary.OutputTokens += usage.OutputTokens
-		summary.CacheReadTokens += usage.CacheReadTokens
-		summary.CacheWriteTokens += usage.CacheWriteTokens
-		total := usage.TotalTokens
-		if total == 0 {
-			total = usage.InputTokens + usage.OutputTokens
-		}
-		summary.TotalTokens += total
-	}
-	return summary
-}
-
-func registerUsagePricingFlags(flags *flag.FlagSet) *usagePricing {
-	var pricing usagePricing
-	flags.Float64Var(&pricing.InputUSDPerMillion, "usage-input-price", 0, "USD per 1M input tokens for --usage cost estimates")
-	flags.Float64Var(&pricing.OutputUSDPerMillion, "usage-output-price", 0, "USD per 1M output tokens for --usage cost estimates")
-	flags.Float64Var(&pricing.CacheReadUSDPerMillion, "usage-cache-read-price", 0, "USD per 1M cache-read tokens for --usage cost estimates")
-	flags.Float64Var(&pricing.CacheWriteUSDPerMillion, "usage-cache-write-price", 0, "USD per 1M cache-write tokens for --usage cost estimates")
-	return &pricing
-}
-
-func markUsagePricingFlagState(flags *flag.FlagSet, pricing *usagePricing) {
-	flags.Visit(func(flag *flag.Flag) {
-		switch flag.Name {
-		case "usage-input-price", "usage-output-price", "usage-cache-read-price", "usage-cache-write-price":
-			pricing.Enabled = true
-		}
-	})
-}
-
-func validateUsagePricing(pricing usagePricing) error {
-	for name, value := range map[string]float64{
-		"--usage-input-price":       pricing.InputUSDPerMillion,
-		"--usage-output-price":      pricing.OutputUSDPerMillion,
-		"--usage-cache-read-price":  pricing.CacheReadUSDPerMillion,
-		"--usage-cache-write-price": pricing.CacheWriteUSDPerMillion,
-	} {
-		if value < 0 {
-			return fmt.Errorf("%s must be non-negative", name)
-		}
-	}
-	return nil
-}
-
-func usagePricingEnabled(pricing usagePricing) bool {
-	return pricing.Enabled
-}
-
-func estimateUsageCostUSD(summary usageSummary, pricing usagePricing) float64 {
-	const tokensPerMillion = 1_000_000
-	return (float64(summary.InputTokens)*pricing.InputUSDPerMillion +
-		float64(summary.OutputTokens)*pricing.OutputUSDPerMillion +
-		float64(summary.CacheReadTokens)*pricing.CacheReadUSDPerMillion +
-		float64(summary.CacheWriteTokens)*pricing.CacheWriteUSDPerMillion) / tokensPerMillion
-}
-
-func writeUsageSummary(w io.Writer, summary usageSummary, pricing usagePricing) {
-	if !summary.HasUsage {
-		return
-	}
-	parts := []string{
-		fmt.Sprintf("input=%d", summary.InputTokens),
-		fmt.Sprintf("output=%d", summary.OutputTokens),
-	}
-	if summary.CacheReadTokens != 0 {
-		parts = append(parts, fmt.Sprintf("cache_read=%d", summary.CacheReadTokens))
-	}
-	if summary.CacheWriteTokens != 0 {
-		parts = append(parts, fmt.Sprintf("cache_write=%d", summary.CacheWriteTokens))
-	}
-	parts = append(parts, fmt.Sprintf("total=%d", summary.TotalTokens))
-	if usagePricingEnabled(pricing) {
-		parts = append(parts, fmt.Sprintf("cost_usd=%.6f", estimateUsageCostUSD(summary, pricing)))
-	}
-	fmt.Fprintf(w, "usage: %s\n", strings.Join(parts, " "))
-}
-
-func buildCodingTools(cfg codingFlagConfig) ([]glue.Tool, toolscoding.Options, error) {
-	return toolscoding.Tools(toolscoding.Options{
-		Enabled:         cfg.Enabled,
-		WorkDir:         cfg.WorkDir,
-		AllowedBinaries: append([]string(nil), cfg.AllowedBinaries...),
-		AllowOverwrite:  cfg.AllowOverwrite,
-	})
-}
-
-// yoloPermission auto-approves every side-effecting tool call. Used by
-// --yolo. The decision is marked RememberFor: RememberSession so a
-// downstream consumer of the daemon protocol that persists remembers
-// records the policy coherently (a yolo run trusts every action; that
-// trust applies for the rest of the session, not forever).
-type yoloPermission struct{}
-
-func (yoloPermission) Decide(_ context.Context, _ glue.PermissionRequest) (glue.PermissionDecision, error) {
-	return glue.PermissionDecision{
-		Allow:       true,
-		Reason:      "auto-approved by --yolo",
-		RememberFor: glue.RememberSession,
-	}, nil
-}
-
-type localPromptPermission struct {
-	input         *bufio.Reader
-	stderr        io.Writer
-	sessionAllows map[string]struct{}
-	targetAllows  map[string]struct{}
-	foreverAllows map[string]struct{}
-}
-
-func newLocalPromptPermission(stdin io.Reader, stderr io.Writer) *localPromptPermission {
-	if stdin == nil {
-		stdin = strings.NewReader("")
-	}
-	if stderr == nil {
-		stderr = io.Discard
-	}
-	return &localPromptPermission{
-		input:         bufio.NewReader(stdin),
-		stderr:        stderr,
-		sessionAllows: map[string]struct{}{},
-		targetAllows:  map[string]struct{}{},
-		foreverAllows: map[string]struct{}{},
-	}
-}
-
-func (p *localPromptPermission) Decide(ctx context.Context, req glue.PermissionRequest) (glue.PermissionDecision, error) {
-	if p == nil {
-		return glue.PermissionDecision{Allow: false, Reason: "permission denied: no local permission handler"}, nil
-	}
-	if _, ok := p.foreverAllows[localPermissionForeverKey(req)]; ok {
-		return glue.PermissionDecision{Allow: true, RememberFor: glue.RememberForever}, nil
-	}
-	if _, ok := p.sessionAllows[localPermissionSessionKey(req)]; ok {
-		return glue.PermissionDecision{Allow: true, RememberFor: glue.RememberSession}, nil
-	}
-	if _, ok := p.targetAllows[localPermissionTargetKey(req)]; ok {
-		return glue.PermissionDecision{Allow: true, RememberFor: glue.RememberSessionTarget}, nil
-	}
-
-	select {
-	case <-ctx.Done():
-		return glue.PermissionDecision{}, ctx.Err()
-	default:
-	}
-	decision, err := promptPermission(p.input, p.stderr, connectPermissionPayload{Request: req})
-	if err != nil {
-		return glue.PermissionDecision{}, err
-	}
-	out := glue.PermissionDecision{Allow: decision.Allow, Reason: decision.Reason}
-	switch decision.RememberFor {
-	case "session":
-		out.RememberFor = glue.RememberSession
-		p.sessionAllows[localPermissionSessionKey(req)] = struct{}{}
-	case "session_target":
-		out.RememberFor = glue.RememberSessionTarget
-		p.targetAllows[localPermissionTargetKey(req)] = struct{}{}
-	case "forever":
-		out.RememberFor = glue.RememberForever
-		p.foreverAllows[localPermissionForeverKey(req)] = struct{}{}
-	}
-	return out, nil
-}
-
-func localPermissionSessionKey(req glue.PermissionRequest) string {
-	return req.SessionID + "\x00" + req.Tool + "\x00" + req.Action
-}
-
-func localPermissionTargetKey(req glue.PermissionRequest) string {
-	return localPermissionSessionKey(req) + "\x00" + req.Target
-}
-
-func localPermissionForeverKey(req glue.PermissionRequest) string {
-	return req.Tool + "\x00" + req.Action + "\x00" + req.Target
-}
-
-func promptPermission(input *bufio.Reader, stderr io.Writer, perm connectPermissionPayload) (connectPermissionDecision, error) {
-	req := perm.Request
-	fmt.Fprintf(stderr, "\nPermission requested: %s", req.Tool)
-	if req.Action != "" {
-		fmt.Fprintf(stderr, " %s", req.Action)
-	}
-	if req.Target != "" {
-		fmt.Fprintf(stderr, "\nTarget: %s", req.Target)
-	}
-	fmt.Fprint(stderr, "\n[d]eny, [a]llow once, allow for [s]ession, session [t]arget, [f]orever: ")
-	raw, err := input.ReadString('\n')
-	if err != nil && !errors.Is(err, io.EOF) {
-		return connectPermissionDecision{}, err
-	}
-	switch strings.ToLower(strings.TrimSpace(raw)) {
-	case "a", "allow", "allow once", "y", "yes":
-		return connectPermissionDecision{Allow: true}, nil
-	case "s", "session":
-		return connectPermissionDecision{Allow: true, RememberFor: "session"}, nil
-	case "t", "target", "session_target", "session target":
-		return connectPermissionDecision{Allow: true, RememberFor: "session_target"}, nil
-	case "f", "forever":
-		return connectPermissionDecision{Allow: true, RememberFor: "forever"}, nil
-	default:
-		return connectPermissionDecision{Allow: false, Reason: "permission denied by user"}, nil
-	}
-}
-
-func postPermissionDecision(ctx context.Context, cfg connectConfig, runID, permissionID string, decision connectPermissionDecision, client httpDoer) error {
-	var body bytes.Buffer
-	if err := json.NewEncoder(&body).Encode(decision); err != nil {
-		return err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, cfg.BaseURL+"/v1/runs/"+url.PathEscape(runID)+"/permissions/"+url.PathEscape(permissionID)+"/decision", &body)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("X-Glue-Client-ID", cfg.ClientID)
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("daemon permission decision: %s", httpStatusError(resp))
-	}
-	return nil
-}
-
-func cancelDaemonRun(ctx context.Context, cfg connectConfig, runID string, client httpDoer) error {
-	req, err := http.NewRequestWithContext(ctx, http.MethodDelete, cfg.BaseURL+"/v1/runs/"+url.PathEscape(runID), nil)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+cfg.Token)
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusAccepted && resp.StatusCode != http.StatusNotFound {
-		return fmt.Errorf("daemon cancel run: %s", httpStatusError(resp))
-	}
-	return nil
-}
-
 // resolveProvider validates the provider name against the registry and
 // resolves the effective model: an explicit --model wins, otherwise the
 // provider's registry default model is used.
@@ -2714,86 +185,6 @@ func newAgent(newProvider providerFactory, cfg agentConfig) (*glue.Agent, error)
 
 func normalizeModel(model string) string {
 	return strings.TrimPrefix(model, "gemini/")
-}
-
-func serveDaemon(ctx context.Context, cfg serveConfig, handler http.Handler, stdout io.Writer) error {
-	return daemon.ServeLocal(ctx, daemon.LocalConfig{
-		Name:            "glue daemon",
-		ListenAddr:      cfg.ListenAddr,
-		Token:           cfg.Token,
-		TokenSource:     cfg.TokenSource,
-		MetadataPath:    cfg.MetadataPath,
-		ShutdownTimeout: cfg.ShutdownTimeout,
-	}, handler, stdout)
-}
-
-func resolveDaemonToken(flagValue string) (token, source string, err error) {
-	return daemon.ResolveToken(flagValue)
-}
-
-func defaultMetadataPath() string {
-	return daemon.DefaultMetadataPath()
-}
-
-func defaultClientID() string {
-	return fmt.Sprintf("cli:%d", os.Getpid())
-}
-
-func readDaemonMetadata(path string) (daemonMetadata, error) {
-	return daemon.ReadMetadata(path)
-}
-
-func writeDaemonMetadata(path string, meta daemonMetadata) error {
-	return daemon.WriteMetadata(path, meta)
-}
-
-func decodePayload[T any](payload any) (T, error) {
-	var out T
-	data, err := json.Marshal(payload)
-	if err != nil {
-		return out, err
-	}
-	err = json.Unmarshal(data, &out)
-	return out, err
-}
-
-func payloadString(payload any, key string) string {
-	m, ok := payload.(map[string]any)
-	if !ok {
-		return ""
-	}
-	value, _ := m[key].(string)
-	return value
-}
-
-func payloadErrorMessage(payload any) string {
-	m, ok := payload.(map[string]any)
-	if !ok {
-		return ""
-	}
-	errValue, ok := m["error"].(map[string]any)
-	if !ok {
-		return ""
-	}
-	msg, _ := errValue["message"].(string)
-	return msg
-}
-
-func httpStatusError(resp *http.Response) string {
-	var out struct {
-		Error struct {
-			Code    string `json:"code"`
-			Message string `json:"message"`
-		} `json:"error"`
-	}
-	_ = json.NewDecoder(resp.Body).Decode(&out)
-	if out.Error.Message != "" {
-		if out.Error.Code != "" {
-			return fmt.Sprintf("%s: %s", out.Error.Code, out.Error.Message)
-		}
-		return out.Error.Message
-	}
-	return resp.Status
 }
 
 // printVersion writes a short version banner sourced from the linker-
